backend/pyproject.toml
[tool.poetry]
authors = [
    "Josh Bradley <joshbradley@microsoft.com>",
    "Newman Cheng <newmancheng@microsoft.com>",
    "Christine DiFonzo <cdifonzo@microsoft.com>",
    "Gabriel Nieves <gnievesponce@microsoft.com>",
    "Douglas Orbaker <dorbaker@microsoft.com>",
    "Chris Sanchez <chrissanchez@microsoft.com>",
    "Shane Solomon <shane.solomon@microsoft.com>",
    "Kenny Zhang <zhangken@microsoft.com>",
]
description = "A web API wrapper around the official GraphRAG library."
license = "MIT"
name = "graphrag-app"
version = "1.2.0"

[tool.poetry.dependencies]
python = "~3.10"

[tool.poetry.group.dev.dependencies]
detect-secrets = ">=1.5.0"
devtools = ">=0.12.2"
flake8 = ">=6.1.0"
ipython = "*"
jupyter = "*"
pre-commit = ">=3.6.0"
ruff = ">=0.1.13"

[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
pytest = ">=8.2.1"
pytest-asyncio = "^0.25.0"
pytest-cov = "^6.0.0"
pytest-env = "^1.1.5"
pytest-xdist = "^3.6.1"
wikipedia = ">=1.4.0"

[tool.poetry.group.backend.dependencies]
adlfs = ">=2024.7.0"
attrs = ">=23.2.0"
azure-core = ">=1.30.1"
azure-cosmos = ">=4.5.1"
azure-identity = ">=1.15.0"
azure-monitor-opentelemetry = "^1.6.4"
azure-search-documents = ">=11.4.0"
azure-storage-blob = ">=12.19.0"
environs = ">=9.5.0"
fastapi = ">=0.110.0"
fastapi-offline = ">=1.7.3"
fastparquet = ">=2023.10.1"
fsspec = ">=2024.2.0"
graphrag = "==1.2.0"
httpx = ">=0.25.2"
kubernetes = ">=29.0.0"
networkx = ">=3.2.1"
nltk = "*"
pandas = ">=2.2.1"
pyaml-env = ">=1.2.1"
pyarrow = ">=15.0.0"
pydantic = ">=1.10.14"
python-multipart = ">=0.0.6"
requests = "*"
rich = ">=13.7.1"
tiktoken = ">=0.6.0"
urllib3 = ">=2.2.2"
uvicorn = ">=0.23.2"

[tool.ruff]
indent-width = 4
line-length = 88
target-version = "py310"

[tool.ruff.format]
preview = true
quote-style = "double"

[tool.ruff.lint]
ignore = ["E402", "E501", "F821"]
preview = true
select = ["E", "F", "I"]

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]
